Split out subscription config, event worker
The dynamic class construction for event handler workers is about to get more complicated. The configuration for a given subscription was a bit complicated, and just a hash of data with checks and coercions here and there. That's been split out into a PORO, and the event handler worker now has a clear interface from the abstract class.

Also fixed up some basic ruby gem stuff. Re-opening modules instead of MyModule::MyClass is more flexible. Errors should be in their own file. A folder in lib/reactor should be included via a file of the same name - normally, that's because said file defines a namespace for the included modules & classes, but we don't do that here yet.

Made some methods to clear out law-of-demeter violations on the Reactor module.