Make console confirmation optional
Event publishing bombs out in a production console in order to avoid
unintended side effects, thus requiring a confirmation in the `publish`
method to actually execute.

In smaller deployments or deployments where service patterns are used,
`publish` calls may in fact be intended and they may being called
indirectly.

There isn't any configuration interface for Reactor as a gem, so this
commit slides in an environment variable to control this behavior.